requirements.txt
fastapi>=0.95
uvicorn>=0.22
gTTS>=2.3.0
playsound>=1.2.2
pydub>=0.25.1
openai>=0.27.0
soundfile>=0.12.1
python-dotenv>=1.0.0
# Use a released whisper version compatible with pip
whisper>=1.1.10